requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
google-generativeai==0.7.2
python-multipart==0.0.9
